version 0.5-3

+++ R/utils.r
@@ -45,27 +45,31 @@ site.summ <- function(y, max.cut=c(2, 5, 10, 20)) {
 }
 
 write.list.Excel <- function(x, fName) {
-    require(RODBC)
-    if (file.exists(fName)) 
-        if (!file.remove(fName))
-           stop("Could not remove existing file - is it open?")
-    on.exit(odbcCloseAll())
-    if (class(x) != "list")
-       stop("object should be a list")
-    if (require(RODBC)==FALSE) {
-       stop("This function requires package RODBC")
-    }
-    channel <- odbcConnectExcel(fName, readOnly=FALSE)
-    if (channel == -1)
-       stop(paste("Could not open file ", fName, "for writing. Is it already open?", sep=""))
-    nms <- names(x)
-    for (i in 1:length(x)) {
-       if (class(x[[i]]) == "data.frame") {
-          colnames(x[[i]]) <- gsub("\\.", "_", colnames(x[[i]]))
-          sqlSave(channel, x[[i]], nms[i], colnames=FALSE, rownames=TRUE)
-       }
-    }    
-    odbcCloseAll()
+  if(.Platform$OS.type == "windows") {
+     require(RODBC)
+     if (file.exists(fName)) 
+         if (!file.remove(fName))
+            stop("Could not remove existing file - is it open?")
+     on.exit(odbcCloseAll())
+     if (class(x) != "list")
+        stop("object should be a list")
+     if (require(RODBC)==FALSE) {
+        stop("This function requires package RODBC")
+     }
+     channel <- odbcConnectExcel(fName, readOnly=FALSE)
+     if (channel == -1)
+        stop(paste("Could not open file ", fName, "for writing. Is it already open?", sep=""))
+     nms <- names(x)
+     for (i in 1:length(x)) {
+        if (class(x[[i]]) == "data.frame") {
+           colnames(x[[i]]) <- gsub("\\.", "_", colnames(x[[i]]))
+           sqlSave(channel, x[[i]], nms[i], colnames=FALSE, rownames=TRUE)
+        }
+     }    
+     odbcCloseAll()
+  } else {
+    stop("This function is only available on Windows. See package WriteXLS for a Linux / MacOS X alternative.")
+  }
 }
 
 Hill.N2 <- function(df, margin=2) {
